Apparently, Virtuoso forcefully casts xsd:int into xsd:integer if you try to insret with SPARQL via Conductor. So better to always insert cr:byteSize always as xsd:integer, as it is anyway a super-type for bith xsd:int and xsd:long. Another issue is that Virtuoso-Sesame driver wrongly returns xsd:int even if it's xsd:integer in the triplestore. So taking this into account in unit tests. Refs #19332.

+++ src/main/java/eionet/cr/dto/ObjectDTO.java
@@ -419,7 +419,7 @@ public class ObjectDTO implements Serializable {
      * @return The literal object.
      */
     public static ObjectDTO createLiteral(int value) {
-        return createLiteral(Integer.valueOf(value), XMLSchema.INT);
+        return createLiteral(Integer.valueOf(value), XMLSchema.INTEGER);
     }
 
     /**
@@ -429,7 +429,7 @@ public class ObjectDTO implements Serializable {
      * @return The literal object.
      */
     public static ObjectDTO createLiteral(long value) {
-        return createLiteral(Long.valueOf(value), XMLSchema.LONG);
+        return createLiteral(Long.valueOf(value), XMLSchema.INTEGER);
     }
 
     /**
